Work Everywhere
Enable the material manager to work on all tabs. The sync icon is now
gray when not in an active glowforge tab. Any temporary materials are
save when the extension window closes thus allowing you to look at
settings in a different window. Sync state is now correct upon opening
the tab, it used to wait for 5 seconds before updating. Do not reload
the page upon removing a material when outside of the GFUI tab.